requirements.txt
streamlit
PyPDF2
google-generativeai
mammoth
python-dotenv